Annotations for `ColumnFunctions` (#971)
- Postpone vectorisation until the last step of
  `specialized_environment` before creation of the DAG.
- Adjust signature s.t. `ColumnFunctions` output
  `(Float|Bool|Int)Columns` instead of scalars
- Make sure processed functions have access to other functions defined
  in the module by explicitly passing on globals to the wrapped function

---------

+++ src/_gettsim/sozialversicherung/arbeitslosen/arbeitslosengeld.py
@@ -55,7 +55,7 @@ def monate_verbleibender_anspruchsdauer(
     monate_durchgängigen_bezugs_von_arbeitslosengeld: float,
     anspruchsdauer_nach_alter: ConsecutiveInt1dLookupTableParamValue,
     anspruchsdauer_nach_versicherungspflichtigen_monaten: ConsecutiveInt1dLookupTableParamValue,
-) -> float:
+) -> int:
     """Remaining amount of months of potential unemployment benefit claims."""
     auf_altersbasis = anspruchsdauer_nach_alter.values_to_look_up[
         alter - anspruchsdauer_nach_alter.base_to_subtract
